@@ -16,12 +16,13 @@ likely lead to crashes. This is only for use in JupyterLab.
 """
 import ast
 import asyncio
+import contextlib
 import importlib.util
 import json
 import os
 import threading
 import weakref
-from typing import Any, Optional, Union, cast
+from typing import Any, cast
 from urllib.parse import parse_qs, quote, urlencode, urlparse, urlunparse
 
 import numpy as np
@@ -85,7 +86,7 @@ class IPyLeafletMixin:
     """
 
     JUPYTER_HOST = '127.0.0.1'
-    JUPYTER_PROXY: Union[str, bool] = os.environ.get('LARGE_IMAGE_JUPYTER_PROXY', 'auto')
+    JUPYTER_PROXY: str | bool = os.environ.get('LARGE_IMAGE_JUPYTER_PROXY', 'auto')
 
     _jupyter_server_manager: Any
 
@@ -164,10 +165,10 @@ class Map:
     """
 
     def __init__(
-            self, *, ts: Optional[IPyLeafletMixin] = None,
-            metadata: Optional[dict] = None, url: Optional[str] = None,
-            gc: Optional[Any] = None, id: Optional[str] = None,
-            resource: Optional[str] = None) -> None:
+            self, *, ts: IPyLeafletMixin | None = None,
+            metadata: dict | None = None, url: str | None = None,
+            gc: Any | None = None, id: str | None = None,
+            resource: str | None = None) -> None:
         """
         Specify the large image to be used with the IPyLeaflet Map.  One of (a)
         a tile source, (b) metadata dictionary and tile url, (c) girder client
@@ -185,7 +186,7 @@ class Map:
             on the girder client.
         """
         self._layer = self._map = self._metadata = self._frame_slider = None
-        self._frame_histograms: Optional[dict[int, Any]] = None
+        self._frame_histograms: dict[int, Any] | None = None
         self._ts = ts
         if (not url or not metadata) and gc and (id or resource):
             fileId = None
@@ -196,10 +197,8 @@ class Map:
                         fileId = entry['_id']
                     id = entry['itemId'] if entry.get('_modelType') == 'file' else entry['_id']
             if id:
-                try:
+                with contextlib.suppress(Exception):
                     metadata = gc.get(f'item/{id}/tiles')
-                except Exception:
-                    pass
                 if metadata:
                     url = gc.urlBase + f'item/{id}/tiles' + '/zxy/{z}/{x}/{y}'
                     if metadata.get('geospatial'):
@@ -276,8 +275,8 @@ class Map:
         return layer
 
     def make_map(
-            self, metadata: dict, layer: Optional[Any] = None,
-            center: Optional[tuple[float, float]] = None) -> Any:
+            self, metadata: dict, layer: Any | None = None,
+            center: tuple[float, float] | None = None) -> Any:
         """
         Create an ipyleaflet map given large_image metadata, an optional
         ipyleaflet layer, and the center of the tile source.
@@ -455,10 +454,10 @@ class Map:
         return JSONDict(self._metadata)
 
     @property
-    def id(self) -> Optional[str]:
+    def id(self) -> str | None:
         return getattr(self, '_id', None)
 
-    def to_map(self, coordinate: Union[list[float], tuple[float, float]]) -> tuple[float, float]:
+    def to_map(self, coordinate: list[float] | tuple[float, float]) -> tuple[float, float]:
         """
         Convert a coordinate from the image or projected image space to the map
         space.
@@ -478,7 +477,7 @@ class Map:
             return tuple(transf.transform(x, y)[::-1])
         return self._metadata['sizeY'] - y, x
 
-    def from_map(self, coordinate: Union[list[float], tuple[float, float]]) -> tuple[float, float]:
+    def from_map(self, coordinate: list[float] | tuple[float, float]) -> tuple[float, float]:
         """
         :param coordinate: a two-tuple that is in the map space coordinates.
         :returns: a two-tuple that is x, y in pixel space or x, y in image
@@ -535,13 +534,12 @@ class Map:
             async def fetch(url):
                 async with aiohttp.ClientSession(
                     timeout=aiohttp.ClientTimeout(total=900),
-                ) as session:
-                    async with session.get(url) as response:
-                        self._frame_histograms[frame] = await response.json()  # type: ignore
-                        # rewrite whole object for watcher
-                        self.frame_selector.frameHistograms = (
-                            self._frame_histograms.copy()  # type: ignore
-                        )
+                ) as session, session.get(url) as response:
+                    self._frame_histograms[frame] = await response.json()  # type: ignore
+                    # rewrite whole object for watcher
+                    self.frame_selector.frameHistograms = (
+                        self._frame_histograms.copy()  # type: ignore
+                    )
 
             asyncio.ensure_future(fetch(histogram_url))
 
